[ana] atlas change to int32 format

+++ libs/mapping/atlas.hpp
@@ -5,8 +5,8 @@
 #include <string>
 class atlas{
 private:
-    image::basic_image<int64_t,3> I;
-    std::vector<int64_t> label_num;
+    image::basic_image<int,3> I;
+    std::vector<int> label_num;
     std::vector<std::string> labels;
     image::matrix<4,4,float> transform;
     bool is_bit_labeled;
@@ -28,7 +28,7 @@ public:
         }
         return labels;
     }
-    const std::vector<int64_t>& get_num(void)
+    const std::vector<int>& get_num(void)
     {
         if(labels.empty())
         {
@@ -38,10 +38,10 @@ public:
         }
         return label_num;
     }
-    int64_t get_label_at(const image::vector<3,float>& mni_space);
+    int get_label_at(const image::vector<3,float>& mni_space);
     std::string get_label_name_at(const image::vector<3,float>& mni_space);
     bool is_labeled_as(const image::vector<3,float>& mni_space,unsigned int label);
-    bool label_matched(int64_t image_label,unsigned int region_label);
+    bool label_matched(int image_label,unsigned int region_label);
 };
 
 #endif // ATLAS_HPP
